requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.6
djangorestframework==3.13.1
idna==3.3
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
Pillow==9.2.0
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.10
